[#43]; feat: v1 api 삭제
