+++ src/main/kotlin/com/yourssu/soongpt/domain/course/business/CourseService.kt
@@ -1,73 +1,47 @@
 package com.yourssu.soongpt.domain.course.business
 
-import com.yourssu.soongpt.domain.course.business.dto.CourseResponse
+import com.yourssu.soongpt.domain.course.business.command.GeneralRequiredCourseFoundCommand
+import com.yourssu.soongpt.domain.course.business.command.MajorElectiveCourseFoundCommand
+import com.yourssu.soongpt.domain.course.business.command.MajorRequiredCourseFoundCommand
+import com.yourssu.soongpt.domain.course.business.command.SearchCoursesCommand
+import com.yourssu.soongpt.domain.course.business.dto.GeneralRequiredCourseResponse
+import com.yourssu.soongpt.domain.course.business.dto.MajorElectiveCourseResponse
+import com.yourssu.soongpt.domain.course.business.dto.MajorRequiredCourseResponse
+import com.yourssu.soongpt.domain.course.business.dto.SearchCoursesResponse
 import com.yourssu.soongpt.domain.course.implement.CourseReader
-import com.yourssu.soongpt.domain.courseTime.implement.CourseTimeReader
 import com.yourssu.soongpt.domain.department.implement.DepartmentReader
 import com.yourssu.soongpt.domain.departmentGrade.implement.DepartmentGradeReader
-import com.yourssu.soongpt.domain.target.implement.TargetReader
-import org.springframework.cache.annotation.Cacheable
 import org.springframework.stereotype.Service
 
 @Service
 class CourseService(
     private val courseReader: CourseReader,
-    private val courseTimeReader: CourseTimeReader,
     private val departmentReader: DepartmentReader,
-    private val targetReader: TargetReader,
     private val departmentGradeReader: DepartmentGradeReader,
 ) {
-    @Cacheable(value = ["courseCache"], key = "'majorRequired'.concat(#command.departmentName).concat(#command.grade)")
-    fun findByDepartmentNameInMajorRequired(command: FoundDepartmentCommand): List<CourseResponse> {
+    fun findAllByDepartmentNameAndGrade(command: MajorRequiredCourseFoundCommand): List<MajorRequiredCourseResponse> {
         val department = departmentReader.getByName(command.departmentName)
         val departmentGrade = departmentGradeReader.getByDepartmentIdAndGrade(department.id!!, command.grade)
-        val courses = courseReader.findAllByDepartmentGradeIdInMajorRequired(departmentGrade.id!!)
-        return filterNotEmptyCourseTimes(courses.map {
-            val courseTimes = courseTimeReader.findAllByCourseId(it.id!!)
-            CourseResponse.from(
-                course = it,
-                target = listOf(targetReader.formatTargetDisplayName(department, departmentGrade)),
-                courseTimes = courseTimes
-            )
-        })
+        val courses = courseReader.findAllByDepartmentGradeInMajorRequired(department, departmentGrade)
+        return courses.map { MajorRequiredCourseResponse.from(it) }
     }
 
-    @Cacheable(value = ["courseCache"], key = "'MajorElective'.concat(#command.departmentName).concat(#command.grade)")
-    fun findByDepartmentNameInMajorElective(command: FoundDepartmentCommand): List<CourseResponse> {
-        val department = departmentReader.getByName(command.departmentName)
-        val courses = courseReader.findAllByDepartmentIdInMajorElective(department.id!!)
-        return filterNotEmptyCourseTimes(courses.map { (course, departmentGrades) ->
-            val courseTimes = courseTimeReader.findAllByCourseId(course.id!!)
-            CourseResponse.from(
-                course = course,
-                target = departmentGrades.map { targetReader.formatTargetDisplayName(department, it) },
-                courseTimes = courseTimes)
-        })
-    }
-
-    @Cacheable(value = ["courseCache"], key = "'GeneralRequired'.concat(#command.departmentName).concat(#command.grade)")
-    fun findByDepartmentNameInGeneralRequired(command: FoundDepartmentCommand): List<CourseResponse> {
+    fun findAllByDepartmentNameAndGrade(command: MajorElectiveCourseFoundCommand): List<MajorElectiveCourseResponse> {
         val department = departmentReader.getByName(command.departmentName)
         val departmentGrade = departmentGradeReader.getByDepartmentIdAndGrade(department.id!!, command.grade)
-        val courses = courseReader.findAllByDepartmentGradeIdInGeneralRequired(departmentGrade.id!!)
-        return filterNotEmptyCourseTimes(courses.map {
-            val courseTimes = courseTimeReader.findAllByCourseId(it.id!!)
-            CourseResponse.from(
-                course = it,
-                target = listOf(targetReader.formatTargetDisplayName(department, departmentGrade)),
-                courseTimes = courseTimes
-            )
-        })
+        val courses = courseReader.findAllByDepartmentGradeInMajorElective(department, departmentGrade)
+        return courses.map { MajorElectiveCourseResponse.from(it) }
     }
 
-    private fun filterNotEmptyCourseTimes(courses: List<CourseResponse>): List<CourseResponse> {
-        return courses.filter { it.courseTime.isNotEmpty() }
+    fun findAllByDepartmentNameAndGrade(command: GeneralRequiredCourseFoundCommand): List<GeneralRequiredCourseResponse> {
+        val department = departmentReader.getByName(command.departmentName)
+        val departmentGrade = departmentGradeReader.getByDepartmentIdAndGrade(department.id!!, command.grade)
+        val courses = courseReader.findAllByDepartmentGradeInGeneralRequired(department, departmentGrade)
+        return courses.map { GeneralRequiredCourseResponse.from(it) }
     }
 
-    fun findById(courseId: Long): CourseResponse {
-        val course = courseReader.getById(courseId)
-        val courseTimes = courseTimeReader.findAllByCourseId(courseId)
-        val target = targetReader.findAllByCourseId(courseId)
-        return CourseResponse.from(course, target, courseTimes)
+    fun search(command: SearchCoursesCommand): SearchCoursesResponse {
+        val courses = courseReader.search(command.query(), command.toPageable())
+        return SearchCoursesResponse.from(courses.content, courses.toPageableInfo())
     }
 }
