+++ src/main/kotlin/com/yourssu/soongpt/domain/course/business/dto/GeneralRequiredCourseResponse.kt
@@ -1,6 +1,6 @@
 package com.yourssu.soongpt.domain.course.business.dto
 
-import com.yourssu.soongpt.domain.course.implement.Course2
+import com.yourssu.soongpt.domain.course.implement.Course
 import com.yourssu.soongpt.domain.courseTime.business.dto.CourseTimeResponse
 
 data class GeneralRequiredCourseResponse(
@@ -15,7 +15,7 @@ data class GeneralRequiredCourseResponse(
 ) {
     companion object {
         fun from(
-            course: Course2,
+            course: Course,
         ): GeneralRequiredCourseResponse {
             return GeneralRequiredCourseResponse(
                 name = course.courseName,
