+++ src/main/kotlin/com/yourssu/soongpt/domain/course/business/dto/SearchCourseResponse.kt
@@ -1,6 +1,6 @@
 package com.yourssu.soongpt.domain.course.business.dto
 
-import com.yourssu.soongpt.domain.course.implement.Course2
+import com.yourssu.soongpt.domain.course.implement.Course
 import com.yourssu.soongpt.domain.courseTime.business.dto.CourseTimeResponse
 
 data class SearchCourseResponse(
@@ -15,7 +15,7 @@ data class SearchCourseResponse(
 ) {
     companion object {
         fun from(
-            course: Course2,
+            course: Course,
         ): SearchCourseResponse {
             return SearchCourseResponse(
                 name = course.courseName,
