+++ src/main/kotlin/com/yourssu/soongpt/domain/course/business/dto/SearchCoursesResponse.kt
@@ -1,6 +1,6 @@
 package com.yourssu.soongpt.domain.course.business.dto
 
-import com.yourssu.soongpt.domain.course.implement.Course2
+import com.yourssu.soongpt.domain.course.implement.Course
 import com.yourssu.soongpt.domain.course.implement.dto.PageableInfo
 
 data class SearchCoursesResponse(
@@ -12,7 +12,7 @@ data class SearchCoursesResponse(
 ) {
     companion object {
         fun from(
-            courses: List<Course2>,
+            courses: List<Course>,
             pageableInfo: PageableInfo,
         ): SearchCoursesResponse {
             val content = courses.map { SearchCourseResponse.from(it) }
