+++ src/main/kotlin/com/yourssu/soongpt/domain/course/implement/dto/SearchCourseDto.kt
@@ -1,10 +1,10 @@
 package com.yourssu.soongpt.domain.course.implement.dto
 
-import com.yourssu.soongpt.domain.course.implement.Course2
+import com.yourssu.soongpt.domain.course.implement.Course
 import org.springframework.data.domain.Pageable
 
 data class SearchCourseDto(
-    val content: List<Course2>,
+    val content: List<Course>,
     val totalElements: Long,
     val totalPages: Int,
     val size: Int,
@@ -12,7 +12,7 @@ data class SearchCourseDto(
 ) {
     companion object {
         fun from(
-            courses: List<Course2>,
+            courses: List<Course>,
             totalElements: Long,
             totalPages: Int,
             size: Int,
